requirements.txt
amqp==5.0.6
asgiref==3.5.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
cachetools==4.2.4
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
distlib==0.3.3
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-extra-fields==3.4.0
environ==1.0
filelock==3.3.1
Flask==2.0.2
Flask-Cors==3.0.10
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.1
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.2
MarkupSafe==2.1.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.0
pipenv==2021.11.15
platformdirs==2.4.0
prompt-toolkit==3.0.22
protobuf==3.18.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
twilio==7.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==6.0.0
